+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog Déboucled
 
+## [2.76.0] (30/03/2026)
+- Réactions emoji et stickers sur les messages du forum (Rand0max)
+- Option pour activer/désactiver les réactions du forum (Rand0max)
+- Option pour activer/désactiver les réactions du chat Décensured (Rand0max)
+
 ## [2.75.0] (30/03/2026)
 
 - Réactions emoji et stickers sur les messages du Chat Décensured (Rand0max)